rename Notification to Toast
to not conflict with the global Notification type
