+++ frontend/components/toast.tsx
@@ -1,10 +1,10 @@
 import { useCallback } from 'react';
-import { Notification, useSelector, useDispatch } from '../store';
+import { ToastData, useSelector, useDispatch } from '../store';
 
 import './notification.css';
 
-export function Notification() {
-  const notification = useSelector(s => s.notification || null);
+export function Toast() {
+  const notification = useSelector(s => s.toast || null);
   return notification && (
     <div className="notification" data-type={notification.type}>
       <div className="notification__content">{notification.text}</div>
@@ -14,13 +14,13 @@ export function Notification() {
 
 const NOTIFICATION_DURATION = 3000;
 
-export function useNotification() {
+export function useToast() {
   const dispatch = useDispatch();
 
-  const notify = useCallback((notification: Notification) => {
-    dispatch({ type: 'set-notification', payload: notification });
+  const notify = useCallback((data: ToastData) => {
+    dispatch({ type: 'set-toast', payload: data });
     setTimeout(() => {
-      dispatch({ type: 'set-notification', payload: undefined });
+      dispatch({ type: 'set-toast', payload: undefined });
     }, NOTIFICATION_DURATION);
   }, [dispatch]);
